% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtex_reader.R
\name{bibtex_reader}
\alias{bibtex_reader}
\title{bibtex reader}
\usage{
bibtex_reader(x)
}
\arguments{
\item{x}{a \code{handlr} internal format object}
}
\value{
an object of class \code{BibEntry}
}
\description{
bibtex reader
}
\examples{
(z <- system.file('extdata/crossref.bib', package = "handlr"))
bibtex_reader(x = z)
(z <- system.file('extdata/bibtex.bib', package = "handlr"))
bibtex_reader(x = z)
}
\seealso{
Other readers: \code{\link{citeproc_reader}},
  \code{\link{codemeta_reader}}, \code{\link{ris_reader}}

Other bibtex: \code{\link{bibtex_writer}}
}
\concept{bibtex}
\concept{readers}
