% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtex_writer.R
\name{bibtex_writer}
\alias{bibtex_writer}
\title{citeproc reader}
\usage{
bibtex_writer(z)
}
\arguments{
\item{file}{(character) a file path}
}
\value{
xxx
}
\description{
citeproc reader
}
\examples{
z <- system.file('extdata/citeproc.json', package = "handlr")
tmp <- citeproc_reader(file = z)
bibtex_writer(tmp)
}
