% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c-handl.R
\name{c.handl}
\alias{c.handl}
\title{combine many handl objects}
\usage{
\method{c}{handl}(...)
}
\arguments{
\item{...}{one or more objects of class \code{handl}; see \link{handl} for more}
}
\value{
text if one RIS citation or list of many
}
\description{
combine many handl objects
}
\examples{
z <- system.file('extdata/crossref.ris', package = "handlr")
cr <- ris_reader(z)
z <- system.file('extdata/peerj.ris', package = "handlr")
prj <- ris_reader(z)
c(cr, prj)
}
