% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff_reader.R
\name{cff_reader}
\alias{cff_reader}
\title{Citation File Format (cff) reader}
\usage{
cff_reader(x)
}
\arguments{
\item{x}{(character) a file path or a yaml string}
}
\value{
an object of class \code{handl}; see \link{handl} for more
}
\description{
Citation File Format (cff) reader
}
\examples{
(z <- system.file('extdata/citation.cff', package = "handlr"))
cff_reader(x = z)
}
\references{
CFF format:
https://github.com/citation-file-format/citation-file-format/blob/master/README.md
}
\seealso{
Other readers: 
\code{\link{bibtex_reader}()},
\code{\link{citeproc_reader}()},
\code{\link{codemeta_reader}()},
\code{\link{ris_reader}()}
}
\concept{cff}
\concept{readers}
