% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_types.R
\docType{data}
\name{cff_reference_types}
\alias{cff_reference_types}
\title{cff references types}
\format{
An object of class \code{character} of length 47.
}
\usage{
cff_reference_types
}
\description{
cff references types
}
\details{
cff citation format types for references
}
\references{
https://bit.ly/2PRK1Vt
}
\keyword{datasets}
