% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff_writer.R
\name{cff_writer}
\alias{cff_writer}
\title{Citation File Format (cff) writer}
\usage{
cff_writer(z, path = NULL)
}
\arguments{
\item{z}{an object of class \code{handl}; see \link{handl} for more}

\item{path}{a file path or connection; default: \code{stdout()}}
}
\value{
text if one cff citation or list of many
}
\description{
Citation File Format (cff) writer
}
\details{
uses \code{yaml::write_yaml} to write to yaml format that
CFF uses
}
\examples{
(z <- system.file('extdata/citation.cff', package = "handlr"))
res <- cff_reader(x = z)
res
unclass(res)
cff_writer(res)
cat(cff_writer(res))
f <- tempfile()
cff_writer(res, f)
readLines(f)
unlink(f)
}
\references{
CFF format:
https://github.com/citation-file-format/citation-file-format
}
\seealso{
Other writers: 
\code{\link{bibtex_writer}()},
\code{\link{citeproc_writer}()},
\code{\link{codemeta_writer}()},
\code{\link{rdf_xml_writer}()},
\code{\link{ris_writer}()},
\code{\link{schema_org_writer}()}

Other cff: 
\code{\link{cff_reader}()}
}
\concept{cff}
\concept{writers}
