% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citeproc_reader.R
\name{citeproc_reader}
\alias{citeproc_reader}
\title{citeproc reader}
\usage{
citeproc_reader(file)
}
\arguments{
\item{file}{(character) a file path}
}
\value{
xxx
}
\description{
citeproc reader
}
\examples{
z <- system.file('extdata/citeproc.json', package = "handlr")
citeproc_reader(file = z)
}
\seealso{
Other readers: \code{\link{bibtex_reader}},
  \code{\link{ris_reader}}

Other citeproc: \code{\link{citeproc_writer}},
  \code{\link{ris_reader}}
}
\concept{citeproc}
\concept{readers}
