% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citeproc_writer.R
\name{citeproc_writer}
\alias{citeproc_writer}
\title{citeproc writer}
\usage{
citeproc_writer(z)
}
\arguments{
\item{z}{a \code{handlr} internal format object}
}
\value{
xxx
}
\description{
citeproc writer
}
\examples{
z <- system.file('extdata/citeproc.json', package = "handlr")
(tmp <- citeproc_reader(file = z))
citeproc_writer(z = tmp)
cat(ris_writer(z = tmp))
}
\seealso{
Other writers: \code{\link{bibtex_writer}},
  \code{\link{ris_writer}}

Other citeproc: \code{\link{citeproc_reader}},
  \code{\link{ris_reader}}
}
\concept{citeproc}
\concept{writers}
