% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codemeta_reader.R
\name{codemeta_reader}
\alias{codemeta_reader}
\title{codemeta reader}
\usage{
codemeta_reader(x)
}
\arguments{
\item{x}{a \code{handlr} internal format object}
}
\value{
an object of class \code{list}
}
\description{
codemeta reader
}
\examples{
(z <- system.file('extdata/codemeta.json', package = "handlr"))
codemeta_reader(x = z)
}
\seealso{
Other readers: \code{\link{bibtex_reader}},
  \code{\link{citeproc_reader}}, \code{\link{ris_reader}}

Other codemeta: \code{\link{codemeta_writer}}
}
\concept{codemeta}
\concept{readers}
