% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{endpoint_open}
\alias{endpoint_open}
\title{Check if endpoint is open}
\usage{
endpoint_open(endpoint, open = "all")
}
\arguments{
\item{endpoint}{Named list with \code{pkg} and \code{fun} strings.}

\item{open}{Endpoints that are open to non-authenticated users.
Specify as named list of vectors.
for example: list(package_name = 'package_function')}
}
\value{

}
\description{
Check if endpoint is open
}
