% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{endpoint_valid}
\alias{endpoint_valid}
\title{Check if endpoint is allowed/exists}
\usage{
endpoint_valid(endpoint, packages)
}
\arguments{
\item{endpoint}{Character vector length two: package and function name}

\item{allowed_packages}{Character vector of allowed package endpoints}
}
\value{

}
\description{
Check if endpoint is allowed/exists
}
