% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{errors_are_false}
\alias{errors_are_false}
\title{Convert errors to FALSE.}
\usage{
errors_are_false(exp)
}
\arguments{
\item{exp}{expresion. The expression to run.}
}
\description{
Convert errors to FALSE.
}

