% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{errors_are_warnings}
\alias{errors_are_warnings}
\title{Convert errors to warnings.}
\usage{
errors_are_warnings(exp)
}
\arguments{
\item{exp}{expresion. The expression to run.}
}
\description{
Convert errors to warnings.
}

