% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{get_endpoint}
\alias{get_endpoint}
\title{Gets the package and function names from a path}
\usage{
get_endpoint(SERVER)
}
\arguments{
\item{SERVER}{rApache variable: list}
}
\value{
character vector of length 2 with first string being the package name and the second the function name
}
\description{
Gets the package and function names from a path
}
\keyword{internal}
