% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{get_env}
\alias{get_env}
\title{Get environment variable.}
\usage{
get_env(var_name)
}
\arguments{
\item{var_name}{Environment variable to get.}
}
\value{
Value of environment variable.
}
\description{
Throws error is environment variable is not defined.
}
