% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{get_exports}
\alias{get_exports}
\title{Reads the NAMESPACE file for installed package}
\usage{
get_exports(package)
}
\arguments{
\item{package}{}
}
\value{

}
\description{
Reads the NAMESPACE file for installed package
}
