% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{get_fun}
\alias{get_fun}
\title{Gets the first relative endpoint (function name) from a path}
\usage{
get_fun(SERVER)
}
\arguments{
\item{SERVER}{rApache variable: list}
}
\value{
endoint character
}
\description{
Gets the first relative endpoint (function name) from a path
}
\keyword{internal}
