% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handl.R
\name{handl}
\alias{handl}
\title{handl object}
\description{
handl object
}
\details{
a handl object is what's returned from the reader functions,
and looks something like:\preformatted{<handl>
  from: codemeta
  many: TRUE
  count: 2
  first 10
    id/doi: https://doi.org/10.5063\%2ff1m61h5x
    id/doi: https://doi.org/10.5063\%2ff1m61h5x
}

Details on each entry:
\itemize{
\item from: the data type the citations come from
\item many: is there more than 1 citation?
\item count: number of citations
\item finally, some details of the first 10 are printed
}
}
