% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{handle}
\alias{handle}
\title{Handles incomming rApache requests}
\usage{
handle(SERVER, packages, open = "all", secret = NULL, timeout = 0,
  rlimits = NULL)
}
\arguments{
\item{SERVER}{rApache variable.}

\item{packages}{Vector of packages that api allows.}

\item{open}{Endpoints that are open to non-authenticated users.
Specify as list of length 2 character vectors
for example: list(c('package_name', 'package_function'))}

\item{timeout}{Time limit in seconds for function to execute.}

\item{rlimits}{named vector/list with rlimit values,
for example: c(cpu = 60, fsize = 1e6).}

\item{key}{string or raw vector needed to decode JSON web token.}
}
\description{
Handles incomming rApache requests
}
