% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlr-package.r
\docType{package}
\name{handlr-package}
\alias{handlr-package}
\alias{handlr}
\title{\strong{Citation format converter}}
\description{
\strong{Citation format converter}
}
\author{
Scott Chamberlain \email{sckott@protonmail.com}
}
