% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{preflight}
\alias{preflight}
\title{Sets CORS headers and checks if preflight}
\usage{
preflight(SERVER)
}
\arguments{
\item{SERVER}{rApache variable}
}
\value{
TRUE if method is OPTIONS otherwise FALSE.
}
\description{
Sets CORS headers and checks if preflight
}
