% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{rapache}
\alias{rapache}
\title{Helper function to permit testing.}
\usage{
rapache(rapache_function, ...)
}
\arguments{
\item{rapache_function}{Name of rApache function to evaluate.}

\item{...}{Additional arguments to rApache function.}
}
\value{

}
\description{
Calls rApache function if it can
}
