% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ris_reader.R
\name{ris_reader}
\alias{ris_reader}
\title{citeproc reader}
\usage{
ris_reader(file)
}
\arguments{
\item{file}{(character) a file path}
}
\value{
xxx
}
\description{
citeproc reader
}
\examples{
z <- system.file('extdata/crossref.ris', package = "handlr")
ris_reader(file = z)

z <- system.file('extdata/peerj.ris', package = "handlr")
ris_reader(file = z)

z <- system.file('extdata/plos.ris', package = "handlr")
ris_reader(file = z)
}
\references{
RIS tags https://en.wikipedia.org/wiki/RIS_(file_format)
}
\seealso{
Other readers: \code{\link{bibtex_reader}},
  \code{\link{citeproc_reader}}

Other citeproc: \code{\link{citeproc_reader}},
  \code{\link{citeproc_writer}}
}
\concept{citeproc}
\concept{readers}
