% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ris_writer.R
\name{ris_writer}
\alias{ris_writer}
\title{ris writer}
\usage{
ris_writer(z)
}
\arguments{
\item{z}{a \code{handlr} internal format object}
}
\value{
xxx
}
\description{
ris writer
}
\examples{
# from a RIS file
z <- system.file('extdata/crossref.ris', package = "handlr")
tmp <- ris_reader(file = z)
cat(ris_writer(z = tmp))

# peerj
z <- system.file('extdata/peerj.ris', package = "handlr")
tmp <- ris_reader(file = z)
cat(ris_writer(z = tmp))

# plos
z <- system.file('extdata/plos.ris', package = "handlr")
tmp <- ris_reader(file = z)
cat(ris_writer(z = tmp))

# elsevier
z <- system.file('extdata/elsevier.ris', package = "handlr")
tmp <- ris_reader(file = z)
cat(ris_writer(z = tmp))

z <- system.file('extdata/citeproc.json', package = "handlr")
res <- citeproc_reader(file = z)
cat(ris_writer(z = res))
}
\references{
RIS tags https://en.wikipedia.org/wiki/RIS_(file_format)
}
\seealso{
Other writers: \code{\link{bibtex_writer}},
  \code{\link{citeproc_writer}},
  \code{\link{codemeta_writer}},
  \code{\link{rdf_xml_writer}},
  \code{\link{schema_org_writer}}

Other ris: \code{\link{ris_reader}}
}
\concept{ris}
\concept{writers}
