% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_org_writer.R
\name{schema_org}
\alias{schema_org}
\alias{schema_hsh}
\title{Schema org writer}
\usage{
schema_hsh(x)

schema_org(x, auto_unbox = TRUE, pretty = TRUE, ...)
}
\arguments{
\item{x}{input}

\item{auto_unbox}{(logical) automatically ‘unbox’ all atomic
vectors of length 1 (default: \code{TRUE}). passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}}

\item{pretty}{(logical) adds indentation whitespace to JSON output
(default: \code{TRUE}), passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}}

\item{...}{further params passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}}
}
\description{
Schema org writer
}
\examples{
(z <- system.file('extdata/bibtex.bib', package = "handlr"))
(tmp <- bibtex_reader(z))
schema_hsh(tmp)
schema_org(tmp)
}
