% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{set_status}
\alias{set_status}
\title{Helper function to permit testing. Instead of setting status directly,
checks if environment is rApache and only calls setStatus if so.}
\usage{
set_status(status, message)
}
\arguments{
\item{message}{}
}
\value{

}
\description{
Helper function to permit testing. Instead of setting status directly,
checks if environment is rApache and only calls setStatus if so.
}
