% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_stack.R
\name{try_stack}
\alias{try_stack}
\title{Evaluate an expression multiple times and if fails, exit R with an exit_code.
This is a wrapper that combines functionality of try_with_exit_code and with_retries.}
\usage{
try_stack(expr, num_tries = 1, sleep_secs = 0.001, exit_code = 1)
}
\arguments{
\item{expr}{expression. The expression to evaluate.}

\item{num_tries}{numeric. The number of attempts.}

\item{exit_code}{numeric. The value of the exit_code to exit R.}

\item{sleep}{numeric. How long to wait between attempts.}
}
\value{
TRUE if the expression gets evaluated successfully; exit R otherwise.
}
\description{
Evaluate an expression multiple times and if fails, exit R with an exit_code.
This is a wrapper that combines functionality of try_with_exit_code and with_retries.
}

