% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_stack.R
\name{try_with_exit_code}
\alias{try_with_exit_code}
\title{Evaluate an expression and exit R with an exit code if expr fails.}
\usage{
try_with_exit_code(expr, exit_code = 1)
}
\arguments{
\item{expr}{expression. The expression to evaluate.}

\item{exit_code}{numeric. The value of the exit_code to exit R.}
}
\value{
TRUE if the expression gets evaluated successfully; exit R otherwise.
}
\description{
Evaluate an expression and exit R with an exit code if expr fails.
}

