% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{valid_endpoint}
\alias{valid_endpoint}
\title{Check if endpoint is allowed/exists}
\usage{
valid_endpoint(endpoint, allowed_packages)
}
\arguments{
\item{endpoint}{Character vector length two: package and function name}

\item{allowed_packages}{Character vector of allowed package endpoints}
}
\value{

}
\description{
Check if endpoint is allowed/exists
}
