% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{validate_endpoint}
\alias{validate_endpoint}
\title{Check if endpoint is allowed/exists}
\usage{
validate_endpoint(endpoint, packages)
}
\arguments{
\item{endpoint}{Named list with \code{pkg} and \code{fun} strings.}

\item{allowed_packages}{Character vector of allowed package endpoints}
}
\value{

}
\description{
Check if endpoint is allowed/exists
}
