% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{with_errors_as}
\alias{errors_are}
\alias{with_errors_as}
\title{Convert errors into another R object.}
\usage{
with_errors_as(are, exp)

errors_are(are, exp)
}
\arguments{
\item{are}{object. The R object to return.}

\item{exp}{expression. The expression to run.}
}
\description{
Instead of erroring, that R object will be returned instead.
}

