% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry.R
\name{with_retries}
\alias{with_retries}
\title{Try evaluating an expression multiple times before erroring.}
\usage{
with_retries(expr, num_tries = 1, sleep = 0.001)
}
\arguments{
\item{expr}{expression. The expression to evaluate with retries.}

\item{num_tries}{numeric. The number of attempts.}

\item{sleep}{numeric. How long to wait between attempts.}
}
\value{
TRUE if the expression gets evaluated successfully on some attempt.
}
\description{
Try evaluating an expression multiple times before erroring.
}

