% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtex_writer.R
\name{bibtex_writer}
\alias{bibtex_writer}
\title{bibtex writer}
\usage{
bibtex_writer(z)
}
\arguments{
\item{z}{a \code{handlr} internal format object}
}
\value{
an object of class \code{BibEntry}
}
\description{
bibtex writer
}
\examples{
(z <- system.file('extdata/citeproc.json', package = "handlr"))
(tmp <- citeproc_reader(file = z))
bibtex_writer(z = tmp)
}
\seealso{
Other writers: \code{\link{citeproc_writer}},
  \code{\link{ris_writer}}

Other bibtex: \code{\link{bibtex_reader}}
}
\concept{bibtex}
\concept{writers}
