% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c-handl.R
\name{c.handl}
\alias{c.handl}
\title{combine many handl objects}
\usage{
\method{c}{handl}(...)
}
\arguments{
\item{...}{one or more objects of class \code{handl}; see \link{handl} for more.
all inputs must be of class \code{handl}. if the first input is not of class
\code{handl}, you will not get back an object of class \code{handl}}
}
\value{
an object of class \code{handl} of length equal to number of
\code{handl} objects passed in
}
\description{
combine many handl objects
}
\examples{
z <- system.file('extdata/crossref.ris', package = "handlr")
cr <- ris_reader(z)
z <- system.file('extdata/peerj.ris', package = "handlr")
prj <- ris_reader(z)
res <- c(cr, prj)
res
invisible(lapply(bibtex_writer(res), cat, sep = "\\n\\n"))
}
