% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff_reader.R
\name{cff_reader}
\alias{cff_reader}
\title{Citation File Format (cff) reader}
\usage{
cff_reader(x)
}
\arguments{
\item{x}{(character) a file path or a yaml string}
}
\value{
an object of class \code{handl}; see \link{handl} for more
}
\description{
Citation File Format (cff) reader
}
\details{
CFF only supports one citation, so \code{many} will always be
\code{FALSE}. You can though have many references in your CFF file
associated with the citation.

\code{references} is an optional component in cff files. If included, we check
the following:
\itemize{
\item each reference must have the 3 required fields: type, authors, title
\item type must be in the allowed set, see \link{cff_reference_types}
\item the elements within authors must each be an entity or person object
https://github.com/citation-file-format/citation-file-format/blob/master/README.md#entity-objects
https://github.com/citation-file-format/citation-file-format/blob/master/README.md#person-objects
\item title must be a string
}
}
\examples{
(z <- system.file('extdata/citation.cff', package = "handlr"))
res <- cff_reader(x = z)
res
res$cff_version
res$b_version
res$message
res$id
res$doi
res$title
res$author
res$references

# no references
(z <- system.file('extdata/citation-norefs.cff', package = "handlr"))
out <- cff_reader(x = z)
out
out$references
}
\references{
CFF format:
https://github.com/citation-file-format/citation-file-format/blob/master/README.md
}
\seealso{
Other readers: 
\code{\link{bibtex_reader}()},
\code{\link{citeproc_reader}()},
\code{\link{codemeta_reader}()},
\code{\link{ris_reader}()}
}
\concept{cff}
\concept{readers}
