% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citeproc_reader.R
\name{citeproc_reader}
\alias{citeproc_reader}
\title{citeproc reader}
\usage{
citeproc_reader(x)
}
\arguments{
\item{x}{(character) a file path or string}
}
\value{
an object of class \code{handl}; see \link{handl} for more
}
\description{
citeproc reader
}
\examples{
# single
z <- system.file('extdata/citeproc.json', package = "handlr")
citeproc_reader(x = z)

# many
z <- system.file('extdata/citeproc-many.json', package = "handlr")
citeproc_reader(x = z)
}
\seealso{
Other readers: \code{\link{bibtex_reader}},
  \code{\link{codemeta_reader}}, \code{\link{ris_reader}}

Other citeproc: \code{\link{citeproc_writer}}
}
\concept{citeproc}
\concept{readers}
