% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codemeta_reader.R
\name{codemeta_reader}
\alias{codemeta_reader}
\title{codemeta reader}
\usage{
codemeta_reader(x)
}
\arguments{
\item{x}{(character) a file path or string (character or json)}
}
\value{
an object of class \code{handl}; see \link{handl} for more
}
\description{
codemeta reader
}
\examples{
# single
(z <- system.file('extdata/codemeta.json', package = "handlr"))
codemeta_reader(x = z)

# many
(z <- system.file('extdata/codemeta-many.json', package = "handlr"))
codemeta_reader(x = z)
}
\seealso{
Other readers: 
\code{\link{bibtex_reader}()},
\code{\link{cff_reader}()},
\code{\link{citeproc_reader}()},
\code{\link{ris_reader}()}

Other codemeta: 
\code{\link{codemeta_writer}()}
}
\concept{codemeta}
\concept{readers}
