% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{get_exports}
\alias{get_exports}
\title{Reads the NAMESPACE file for installed package}
\usage{
get_exports(package)
}
\arguments{
\item{package}{Package name to read NAMESPACE from}
}
\value{
character vector of exports from package
}
\description{
Reads the NAMESPACE file for installed package
}
