% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{get_params}
\alias{get_params}
\title{Puts parameters for endpoint function into a list.}
\usage{
get_params(req_data)
}
\arguments{
\item{req_data}{Result of call to \code{get_req}.}
}
\value{
Named list of parameters suplied in request.
}
\description{
Puts parameters for endpoint function into a list.
}
