% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{get_params}
\alias{get_params}
\title{Get parameters from API call}
\usage{
get_params(GET, POST, SERVER)
}
\arguments{
\item{GET}{rApache variable: list of parameters}

\item{POST}{rApache variable: list of parameters}

\item{SERVER}{rApache variable: list}
}
\value{
params named list
}
\description{
Get parameters from API call
}
\keyword{internal}
