% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{handle}
\alias{handle}
\title{Handles incomming rApache requests}
\usage{
handle(GET, POST, COOKIES, FILES, SERVER, allowed_packages, timeout = 0,
  rlimits = NULL)
}
\arguments{
\item{GET}{rApache variable: list of parameters}

\item{POST}{rApache variable: list of parameters}

\item{COOKIES}{rApache variable: list of cookies}

\item{FILES}{rApache variable:}

\item{SERVER}{rApache variable: list}

\item{allowed_packages}{character vector of allowed packages}

\item{timeout}{maximum time in seconds to allow for call to return}

\item{rlimits}{named vector/list with RLIMIT values, for example: c(cpu = 60, fsize = 1e6)}
}
\description{
Handles incomming rApache requests
}
