% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{handle}
\alias{handle}
\title{Handles incomming rApache requests}
\usage{
handle(SERVER, GET, packages, open = "all", secret = NULL, timeout = 0,
  rlimits = NULL)
}
\arguments{
\item{SERVER}{rApache variable.}

\item{GET}{rApache variable.}

\item{packages}{Vector of packages that api allows.}

\item{open}{Endpoints that are open to non-authenticated users.
Specify as named list of vectors.
for example: list(package_name = 'package_function')}

\item{secret}{string or raw vector needed to encode/decode JSON web token.
Gets passed along to endpoints with argument 'secret'.}

\item{timeout}{Time limit in seconds for function to execute.}

\item{rlimits}{named vector/list with rlimit values,
for example: c(cpu = 60, fsize = 1e6).}
}
\description{
Handles incomming rApache requests
}
