% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlr-package.r
\docType{package}
\name{handlr-package}
\alias{handlr-package}
\alias{handlr}
\title{\strong{Citation format converter}}
\description{
A tool for converting among citation formats
}
\section{supported readers}{

\itemize{
\item citeproc
\item ris
\item bibtex
}
}

\section{supported writers}{

\itemize{
\item citeproc
\item ris
\item bibtex
\item schema.org
\item rdfxml (requires suggested package \code{jsonld})
}
}

\author{
Scott Chamberlain \email{sckott@protonmail.com}
}
