% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlr-package.r
\docType{package}
\name{handlr-package}
\alias{handlr-package}
\alias{handlr}
\title{\strong{Citation format converter}}
\description{
A tool for converting among citation formats
}
\section{supported readers}{

\itemize{
\item citeproc
\item ris
\item bibtex (requires suggested package \code{bibtex})
\item codemeta
\item cff
}
}

\section{supported writers}{

\itemize{
\item citeproc
\item ris
\item bibtex
\item schema.org
\item rdfxml (requires suggested package \code{jsonld})
\item codemeta
\item cff
}
}

\section{links for citation formats}{

\itemize{
\item citeproc: \url{https://en.wikipedia.org/wiki/CiteProc}
\item codemeta: \url{https://codemeta.github.io/}
\item ris: \url{https://en.wikipedia.org/wiki/RIS_(file_format)}
\item bibtex: \url{http://www.bibtex.org/}
\item schema.org: \url{https://schema.org/}
\item rdfxml: \url{https://en.wikipedia.org/wiki/RDF/XML}
\item cff: \url{https://citation-file-format.github.io/}
}
}

\author{
Scott Chamberlain \email{sckott@protonmail.com}
}
