% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\docType{data}
\name{HandlrClient}
\alias{HandlrClient}
\title{handlr client}
\description{
handlr client
}
\details{
\strong{Methods}
\describe{
\item{\code{parse()}}{
parse input
}
}
}
\examples{
\dontrun{
# crosscite
z <- system.file('extdata/crosscite.json', package = "handlr")
(x <- HandlrClient$new(x = z, from ="crosscite"))
x$x
x$proc

# read in citeproc, write out bibtex
z <- system.file('extdata/citeproc.json', package = "handlr")
(x <- HandlrClient$new(x = z, from ="citeproc"))
x$path
x$read()
x$write("bibtex")
}
}
\keyword{datasets}
