% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\docType{data}
\name{HandlrClient}
\alias{HandlrClient}
\title{handlr client}
\description{
handlr client
}
\details{
\strong{Methods}
\describe{
\item{\code{read(format = NULL)}}{
read input
format: optionally specify format, e.g., ris. We attempt to
auto-detect this for you.
}
}
}
\examples{
\dontrun{
# crosscite
z <- system.file('extdata/crosscite.json', package = "handlr")
(x <- HandlrClient$new(x = z))
x$path

# read in citeproc, write out bibtex
z <- system.file('extdata/citeproc.json', package = "handlr")
(x <- HandlrClient$new(x = z))
x$path
x$ext
x$read("citeproc")
x$parsed
x$write("bibtex")
f <- tempfile(fileext = ".bib")
x$write("bibtex", file = f)
readLines(f)
unlink(f)

# read in ris, write out ris
z <- system.file('extdata/peerj.ris', package = "handlr")
(x <- HandlrClient$new(x = z))
x$path
x$format
x$read("ris")
x$parsed
x$write("ris")
cat(x$write("ris"))

# handle strings instead of files
z <- system.file('extdata/citeproc-crossref.json', package = "handlr")
(x <- HandlrClient$new(x = readLines(z)))
x$read("citeproc")
x$parsed
cat(x$write("bibtex"), sep = "\\n")
}
}
\keyword{datasets}
