% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\docType{data}
\name{HandlrClient}
\alias{HandlrClient}
\title{handlr client}
\arguments{
\item{x}{(character) a file path}
}
\description{
handlr client
}
\details{
\strong{Methods}
\describe{
\item{\code{read(format, ...)}}{
read input
format: one of citeproc, ris, bibtex
...: further args to the writer fxn, if any
}
\item{\code{write(format, file = NULL, ...)}}{
write
format: one of citeproc, ris, bibtex, rdfxml
file: a file path, if NULL to stdout
...: further args to the writer fxn, if any
}
}
}
\examples{
\dontrun{
# crosscite
z <- system.file('extdata/crosscite.json', package = "handlr")
(x <- HandlrClient$new(x = z))
x$path

# read in citeproc, write out bibtex
z <- system.file('extdata/citeproc.json', package = "handlr")
(x <- HandlrClient$new(x = z))
x$path
x$ext
x$read("citeproc")
x$parsed
x$write("bibtex")
f <- tempfile(fileext = ".bib")
x$write("bibtex", file = f)
readLines(f)
unlink(f)

# read in ris, write out ris
z <- system.file('extdata/peerj.ris', package = "handlr")
(x <- HandlrClient$new(x = z))
x$path
x$format
x$read("ris")
x$parsed
x$write("ris")
cat(x$write("ris"))

# read in bibtex, write out ris
(z <- system.file('extdata/bibtex.bib', package = "handlr"))
(x <- HandlrClient$new(x = z))
x$path
x$format
x$read("bibtex")
x$parsed
x$write("ris")
cat(x$write("ris"))

# read in bibtex, write out RDF XML
(z <- system.file('extdata/bibtex.bib', package = "handlr"))
(x <- HandlrClient$new(x = z))
x$path
x$format
x$read("bibtex")
x$parsed
x$write("rdfxml")
cat(x$write("rdfxml"))

# handle strings instead of files
z <- system.file('extdata/citeproc-crossref.json', package = "handlr")
(x <- HandlrClient$new(x = readLines(z)))
x$read("citeproc")
x$parsed
cat(x$write("bibtex"), sep = "\\n")
}
}
\keyword{datasets}
