% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{rapache}
\alias{rapache}
\title{Helper function to permit testing.}
\usage{
rapache(rapache_function, arguments = list())
}
\arguments{
\item{arguments}{}
}
\description{
Calls rApache function only if rapache environment
}
