% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_org_writer.R
\name{schema_org}
\alias{schema_org}
\alias{schema_hsh}
\title{Schema org writer}
\usage{
schema_hsh(x)

schema_org(x, auto_unbox = TRUE, pretty = TRUE, ...)
}
\arguments{
\item{x}{input}
}
\description{
Schema org writer
}
\examples{
(z <- system.file('extdata/bibtex.bib', package = "handlr"))
(tmp <- bibtex_reader(z))
schema_hsh(tmp)
schema_org(tmp)
}
