% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{validate_jwt}
\alias{validate_jwt}
\title{Validates JSON web token}
\usage{
validate_jwt(req_headers)
}
\arguments{
\item{req_headers}{List with request headers.}
}
\value{
List with claim.
}
\description{
Error occurs if validation fails.
}
